proxy信息解析方法提取
