+++ bistoury-ui/src/main/java/qunar/tc/bistoury/ui/controller/ConfigController.java
@@ -18,7 +18,6 @@
 package qunar.tc.bistoury.ui.controller;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import com.ning.http.client.AsyncHttpClient;
@@ -27,7 +26,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
-import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -38,6 +36,8 @@ import qunar.tc.bistoury.serverside.configuration.DynamicConfigLoader;
 import qunar.tc.bistoury.serverside.configuration.local.LocalDynamicConfig;
 import qunar.tc.bistoury.serverside.util.ResultHelper;
 import qunar.tc.bistoury.ui.service.ProxyService;
+import qunar.tc.bistoury.ui.util.ProxyInfo;
+import qunar.tc.bistoury.ui.util.ProxyInfoParse;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
@@ -56,8 +56,6 @@ public class ConfigController {
     private static final String COLON = ":";
 
     private static final AsyncHttpClient httpClient = AsyncHttpClientHolder.getInstance();
-
-    private static final Splitter PROXY_SPLITTER = Splitter.on(COLON);
 
     private static final TypeReference<ApiResult<AgentInfo>> AGENT_TYPE_REFERENCE = new TypeReference<ApiResult<AgentInfo>>() {
     };
@@ -94,7 +92,7 @@ public class ConfigController {
 
     private void doGetWebSocketUrl(List<String> result, List<String> proxyWebSocketUrls, final String agentIp) {
         for (String proxyWebSocketUrl : proxyWebSocketUrls) {
-            Optional<ProxyInfo> optional = parseProxyInfo(proxyWebSocketUrl);
+            Optional<ProxyInfo> optional = ProxyInfoParse.parseProxyInfo(proxyWebSocketUrl);
             if (!optional.isPresent()) {
                 continue;
             }
@@ -131,16 +129,6 @@ public class ConfigController {
         return false;
     }
 
-    private Optional<ProxyInfo> parseProxyInfo(String line) {
-        List<String> list = PROXY_SPLITTER.splitToList(line);
-        if (CollectionUtils.isEmpty(list) || list.size() != 3) {
-            return Optional.empty();
-        }
-        final String ip = list.get(0);
-        final int tomcatPort = Integer.valueOf(list.get(1));
-        final int websocketPort = Integer.valueOf(list.get(2));
-        return Optional.of(new ProxyInfo(ip, tomcatPort, websocketPort));
-    }
 
     private static class AgentInfo {
         private String ip;
@@ -154,48 +142,5 @@ public class ConfigController {
         }
     }
 
-    private static class ProxyInfo {
-        private String ip;
-        private int tomcatPort;
-        private int websocketPort;
 
-        public ProxyInfo(String ip, int tomcatPort, int websocketPort) {
-            this.ip = ip;
-            this.tomcatPort = tomcatPort;
-            this.websocketPort = websocketPort;
-        }
-
-        public String getIp() {
-            return ip;
-        }
-
-        public void setIp(String ip) {
-            this.ip = ip;
-        }
-
-        public int getTomcatPort() {
-            return tomcatPort;
-        }
-
-        public void setTomcatPort(int tomcatPort) {
-            this.tomcatPort = tomcatPort;
-        }
-
-        public int getWebsocketPort() {
-            return websocketPort;
-        }
-
-        public void setWebsocketPort(int websocketPort) {
-            this.websocketPort = websocketPort;
-        }
-
-        @Override
-        public String toString() {
-            return "ProxyInfo{" +
-                    "ip='" + ip + '\'' +
-                    ", tomcatPort=" + tomcatPort +
-                    ", websocketPort=" + websocketPort +
-                    '}';
-        }
-    }
 }
